requirements.txt
numpy>=1.21
scipy>=1.7
pandas>=1.3
astropy>=5.0
alerce>=1.0
scikit-learn>=1.0
h5py>=3.4
matplotlib>=3.4
seaborn>=0.12
tqdm>=4.62
pyyaml>=6.0
requests>=2.26
click>=8.0
joblib>=1.1
pyarrow>=7.0
# Recommended for fast ANN (install separately):
# faiss-cpu>=1.7   or   faiss-gpu>=1.7
